changed canvas propertis to floats

+++ raydraw.py
@@ -32,23 +32,23 @@ class scene():
         self.hittable = hittable
 
 
-CANVAS_HEIGHT = 40
-CANVAS_WIDTH = 50
+CANVAS_HEIGHT = 40.0
+CANVAS_WIDTH = 50.0
 
-VIEWPOINT_HEIGHT = 100
-VIEWPOINT_WIDTH = 200
-VIEWPOINT_DIST = 5
+VIEWPOINT_HEIGHT = 100.0
+VIEWPOINT_WIDTH = 200.0
+VIEWPOINT_DIST = 5.0
 
 def getViewPointXYZ(canvasx, canvasy):
-    return [canvasx/ CANVAS_WIDTH * VIEWPOINT_WIDTH - VIEWPOINT_WIDTH /2, canvasy / CANVAS_HEIGHT * VIEWPOINT_HEIGHT * - VIEWPOINT_HEIGHT / 2, VIEWPOINT_DIST]
+    return [ canvasx/ CANVAS_WIDTH * VIEWPOINT_WIDTH - VIEWPOINT_WIDTH /2,float(canvasy / CANVAS_HEIGHT * VIEWPOINT_HEIGHT * - VIEWPOINT_HEIGHT / 2), VIEWPOINT_DIST]
 
 def drawcanvas(cscene, filename):
     f = open(filename, "w")
     f.write("P3 \n")
-    f.write(str(CANVAS_WIDTH) + " " + str(CANVAS_HEIGHT) + "\n")
+    f.write(str(int(CANVAS_WIDTH)) + " " + str(int(CANVAS_HEIGHT)) + "\n")
     f.write("255 \n")
-    for i in range(0, CANVAS_HEIGHT):
-        for j in range(0, CANVAS_WIDTH):
+    for i in range(0, int(CANVAS_HEIGHT)):
+        for j in range(0, int(CANVAS_WIDTH)):
             [x, y, z] = getViewPointXYZ(j, i)
             r = ray(np.array([0,0,0]),np.array([x, y, z]))
             for o in cscene.hittable:
